BLOG SHOW PAGE:
ARTICLE IMAGES

If the image in the article is large: If the image is wider than the the div that contains the article text, resize it so that it is the same width as the div.

If the image in the article is small: If the image is less wide than the div that contains the article text, please align it in the centre of the div.

All images must responsively change size depending on the screen size/orientation.