Grading Task
Task A: ‘Select all’ - please make this into a button. It should also appear on the right side of the screen below ‘Grading?’ and above the first checkbox,not the left side of the screen where it is now.

Task B: ‘Grading?’ - This must remain above the checkboxes.

Task C: This information should all stay on the same line where possible. Consider alternating colours to help differentiate students. Dropdown menu must be easy to select on a small screened device.

Task D: Checkboxes are unsuitable for a mobile screen. Please change these to buttons that toggle. If possible, a stylish red box with a cross in it when unchecked and a green box with a tick in it when checked would be fantastic. I’m open to your suggestions of what you think would be possible and an enjoyable user experience.